Source S3: Refactor config validation
